(MERCURY-182) An error occurred due to instances not being created
This commit fixes an issue caused by the loop invariant used when
waiting for the Viya VM nodes to be created.

In install.sh, the "Checking Viya VMs" loop has an invariant which
is always false, thus the body of the loop never executes and the
loop can exit even when all Viya VMs are not yet created. As the code
following that loop assumes the Viya VMs are created, this causes
an error.

This commit modifies the invariant such that it will not exit the loop
until ALL the Viya VMs are created.

Change-Id: I563a7186a44c5d81b237c2f42466d065e45d6db9