(AWSVIYA-132) Multiple concurrent deployments intermittently fail
This commit reinstates a "retry" around the site.yml playbook execution in
the 03-install step of the Ansible Controller. When running multiple deployments
of the same software order in parallel, failures occasionally occur when downloading
SAS RPMs via yum.

The retry catches these failures and re-runs the site.yml playbook, up to a maximum of
three attempts.