(AWSVIYA-194) Updates for 19w34 changes in SAS Orchestration CLI
As of 19w34, hyphens will no longer be allowed in ansible host group names. This
commit updates this quickstart to handle the new host group names (in which
hyphens are replaced by underscores).

Additionally, this commit adds functionality to this quickstart to accomodate overriding the
common code branch and the location of the SAS orchestration CLI .tgz file, in order to
accomodate the need to test versions that have not yet been promoted to production.

See scripts/process_overrides.sh for information on overriding these parameters.