(AWSVIYA-257) Fix recover_cascontroller.sh
Add AnsibleController to the limit list for the prepare_nodes playbook, otherwise
the new CAS instance can still be in the process of initializing when the playbook
tries to run scripts on it. This is because the "wait_for_viya_vms" task will not
run unless the AnsibleController is in the limit list along with the new CAS
controller.